Move set bound restrict from utils/lp to lp/solver
